requirements.txt
requests
ask_sdk_core
ask_sdk_model
ask_sdk_local_debug
python-dotenv
geopy
pytest
python-lambda-local